+++ io.werval.modules/io.werval.modules.smtp/src/main/java/io/werval/modules/smtp/internal/Transport.java
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.werval.modules.smtp;
+package io.werval.modules.smtp.internal;
 
-/* package */ enum Transport
+public enum Transport
 {
     CLEAR, SSL, TLS
 }